extract logic so it can be used with new command
